How to reinstall packages from a previos version with `nvm`
I always forget to add the packages on the first go, so I do the next version
installation, then I try to reinstall packages without installing a new
version (cause I already installed it), and the command doesn't work.

ToDo: maybe I can do a sort of bash/js script to automatically do the
full upgrade, with previous packages installation. But maybe it would be
better to provide an `nvm` command that just does that.